+++ packages/core/src/components/accordion/components/handles/state-icon.handle.tsx
@@ -3,8 +3,8 @@ import { FunctionalComponent, h } from "@stencil/core";
 import { AccordionSectionState } from "../accordion-section.interfaces";
 
 export const HandleStateIcon: FunctionalComponent<{ state: AccordionSectionState }> = ({ state }) => {
-  if (state === "danger") {
-    return <dso-icon icon="status-danger"></dso-icon>;
+  if (state === "error") {
+    return <dso-icon icon="status-error"></dso-icon>;
   }
 
   if (state === "success") {