+++ packages/react/src/components/accordion/accordion.content.tsx
@@ -176,18 +176,18 @@ export const addonsSections: AccordionSection<JSX.Element>[] = [
   },
   {
     ...section1,
-    state: "danger",
+    state: "error",
     icon: "user-line",
     open: true,
   },
   {
     ...section1,
-    state: "danger",
+    state: "error",
     status: "5 van 8 beantwoord",
   },
   {
     ...section1,
-    state: "danger",
+    state: "error",
     status: "5 van 8 beantwoord",
   },
   {